More stable than ever! Uses external GifCreator lib. Still frame timing issues.

+++ interface/src/Application.h
@@ -28,6 +28,7 @@
 #include <AbstractViewStateInterface.h>
 #include <EntityEditPacketSender.h>
 #include <EntityTreeRenderer.h>
+#include <GifCreator.h>
 #include <input-plugins/KeyboardMouseDevice.h>
 #include <input-plugins/TouchscreenDevice.h>
 #include <OctreeQuery.h>
@@ -60,7 +61,6 @@
 #include "scripting/DialogsManagerScriptingInterface.h"
 #include "ui/ApplicationOverlay.h"
 #include "ui/BandwidthDialog.h"
-#include "ui/Gif.h"
 #include "ui/LodToolsDialog.h"
 #include "ui/LogDialog.h"
 #include "ui/OctreeStatsDialog.h"
@@ -613,7 +613,7 @@ private:
 
     QTimer animatedSnapshotTimer;
     GifWriter _animatedSnapshotGifWriter;
-    uint8_t _currentAnimatedSnapshotFrame { 0 };
+    uint32_t _currentAnimatedSnapshotFrame { 0 };
     QString _animatedSnapshotPath;
 };
